backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart

# LangChain dependencies
langchain-core
langchain-community
langchain-text-splitters
langchain-chroma>=0.1.2
langchain-google-genai

# Google AI dependencies
google-generativeai
protobuf==3.20.3

# Vector store dependencies
chromadb>=0.5.3
duckdb>=1.0.0

# PDF processing
pypdf

# FastAPI dependencies
pydantic